requirements/default.txt
# Django
django==4.0.4

django-admin-sortable2==1.0.3
django-advanced-filters==1.3.0
django-celery-results==2.2.0
django-select2==7.6.1
django-session-timeout==0.0.4
django-widget-tweaks==1.4.8
django-extensions==3.1.3
django-admin-rangefilter

django-celery-beat==2.1.0

djangorestframework==3.12.4
djangorestframework-datatables==0.6.0

redis
filesplit==3.0.2
psycopg2-binary
python-dotenv==0.17.0

sentry-sdk==1.2.0
GitPython==3.1.24
